Move settings selectors to a properly-named file
