+++ src/components/scenes/CreateWalletSelectFiatScene.js
@@ -8,7 +8,7 @@ import { connect } from 'react-redux'
 import { FIAT_COUNTRY } from '../../constants/CountryConstants'
 import * as Constants from '../../constants/indexConstants.js'
 import s from '../../locales/strings.js'
-import { getDefaultFiat } from '../../modules/Settings/selectors.js'
+import { getDefaultFiat } from '../../selectors/SettingsSelectors.js'
 import { type Dispatch, type RootState } from '../../types/reduxTypes.js'
 import type { CreateWalletType, FlatListItem, GuiFiatType } from '../../types/types.js'
 import { getSupportedFiats } from '../../util/utils'
