+++ src/components/scenes/DefaultFiatSettingScene.js
@@ -7,8 +7,8 @@ import { connect } from 'react-redux'
 
 import { setDefaultFiatRequest } from '../../actions/SettingsActions'
 import s from '../../locales/strings.js'
-import { getDefaultFiat } from '../../modules/Settings/selectors.js'
 import Text from '../../modules/UI/components/FormattedText/FormattedText.ui.js'
+import { getDefaultFiat } from '../../selectors/SettingsSelectors.js'
 import { THEME } from '../../theme/variables/airbitz'
 import { type Dispatch, type RootState } from '../../types/reduxTypes.js'
 import type { FlatListItem, GuiFiatType } from '../../types/types.js'
