+++ src/selectors/SettingsSelectors.js
@@ -1,6 +1,6 @@
 // @flow
 
-import { type RootState } from '../../types/reduxTypes.js'
+import { type RootState } from '../types/reduxTypes.js'
 
 export const getDefaultFiat = (state: RootState) => {
   const defaultIsoFiat: string = state.ui.settings.defaultIsoFiat
